+++ Harbor_total/src/main/java/com/example/harbor_total/commute/repository/CommuteRepository.java
@@ -1,0 +1,2 @@
+package com.example.harbor_total.Commute.repository;public class CommuteRepository {
+}
