+++ Harbor_total/src/main/java/com/example/harbor_total/commute/repository/EmployeeRepository.java
@@ -4,6 +4,10 @@ import com.example.harbor_total.domain.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface EmployeeRepository extends JpaRepository<Employee, String> {
+
+
 }
